move: _component 폴더로 리뷰 컴포넌트 파일들 이동
